+++ lib/state_extended.dart
@@ -72,23 +72,26 @@ part 'part12_set_state.dart';
 /// class StateXController: dependOnInheritedWidget(), notifyClients()
 part 'part13_statex_controller.dart';
 
+/// Implemented [Listenable] class
+part 'part14_change_notifier_class.dart';
+
 /// [Listenable] widget builder
-part 'part14_listenable_widget_builder_mixin.dart';
+part 'part15_listenable_widget_builder.dart';
 
 /// For StateXController, setState(), stateOf(), ofState(), firstState, and lastState
-part 'part15_set_state_mixin.dart';
+part 'part16_set_state_mixin.dart';
 
 /// State object's controllerByType(), controllerById(), rootCon
-part 'part16_controllers_by_id_mixin.dart';
+part 'part17_controllers_by_id_mixin.dart';
 
 /// The 'Root' State object: lastContext, dataObject and inDebugMode
-part 'part17_root_state_mixin.dart';
+part 'part18_root_state_mixin.dart';
 
 /// Record an exception in a State object
-part 'part18_record_exception_mixin.dart';
+part 'part19_record_exception_mixin.dart';
 
 /// Supply the Async API: initAsync() and onAsyncError()
-part 'part19_async_ops_mixin.dart';
+part 'part20_async_ops_mixin.dart';
 
 /// A UUID generator, useful for generating unique IDs.
-part 'part20_uuid.dart';
+part 'part21_uuid.dart';